Move mainView to views folder
