+++ cypress/integration/userCanSeeSingleArticle.feature.js
@@ -23,5 +23,6 @@ describe('User can see single article', () => {
     );
     cy.get('[data-testid=date]').should('contain', '2021-05-25');
     cy.get('[data-testid=category]').should('contain', 'Science');
+    cy.get('[data-testid=author]').should('contain', 'Mr. Fake');
   });
 });
